[#noissue] Harden C boundary, shutdown, and resource limits across the agent
Catch exceptions in the C for_each trampolines (unwinding through the
user's C frames is UB), guard NULL header key/value, and replace the
one-past-the-end NUL probe in the C adapters with an always-copy helper.

Join gRPC workers inline when the shutdown joiner thread cannot be
created — unwinding with joinable handles would std::terminate the host
under the very resource pressure that made thread creation fail.

Release the active-span registration in ~SpanImpl/~UnsampledSpan so
spans dropped without EndSpan no longer leak map entries and skew
active-request stats for the process lifetime.

Bound the ping-stream pong wait (30s, then cancel and recycle the
stream), the per-span exception buffer (100), and callstack frames
(128); make parseTraceId's noexcept survivable under OOM via a
function-try-block.

Store Span.EventChunkSize as int so negative values hit the minimum
clamp instead of wrapping to SIZE_MAX and silently disabling chunk
flushing, and clamp LRU cache capacity to >= 1 to keep a zero-size
cache from evicting its just-inserted node and returning freed memory.

+++ src/config.h
@@ -242,7 +242,10 @@ namespace pinpoint {
             size_t queue_size = defaults::SPAN_QUEUE_SIZE;
             int max_event_depth = defaults::SPAN_MAX_EVENT_DEPTH;
             int max_event_sequence = defaults::SPAN_MAX_EVENT_SEQUENCE;
-            size_t event_chunk_size = defaults::SPAN_EVENT_CHUNK_SIZE;
+            // int, not size_t: the parsers produce an int, and a negative
+            // value assigned to size_t would wrap past the minimum check in
+            // Config::check(), silently disabling span-event chunk flushing.
+            int event_chunk_size = defaults::SPAN_EVENT_CHUNK_SIZE;
         } span;
 
         struct {
